requirements.txt
openai
python-dotenv
ipywidgets